+++ examples/python/async_httpx.py
@@ -4,7 +4,7 @@ import asyncio
 
 import smello
 
-smello.init()
+smello.init(server_url="http://localhost:5110")
 
 import httpx