+++ examples/python/demo_httpbin.py
@@ -4,7 +4,7 @@ import time
 
 import smello
 
-smello.init()
+smello.init(server_url="http://localhost:5110")
 
 import requests